backend/requirements.txt
pyramid
SQLAlchemy
psycopg2-binary
alembic
zope.sqlalchemy
pyramid_tm
firebase-admin
PyJWT
